Refactor test class mock logger into MockLoggerMixin
