+++ cumulusci/tasks/tests/test_command.py
@@ -13,23 +13,15 @@ from cumulusci.core.config import BaseGlobalConfig
 from cumulusci.core.config import BaseProjectConfig
 from cumulusci.core.config import OrgConfig
 from cumulusci.core.config import TaskConfig
-from cumulusci.core.tests.utils import MockLoggingHandler
+from cumulusci.core.tests.utils import MockLoggerMixin
 
 from cumulusci.tasks.command import Command
 from cumulusci.tasks.command import SalesforceCommand
 from cumulusci.tasks.command import CommandException
 
 
-class TestCommandTask(unittest.TestCase):
+class TestCommandTask(MockLoggerMixin, unittest.TestCase):
     """ Tests for the basic command task """
-
-    @classmethod
-    def setUpClass(cls):
-        super(TestCommandTask, cls).setUpClass()
-        logger = logging.getLogger(cumulusci.core.tasks.__name__)
-        logger.setLevel(logging.DEBUG)
-        cls._task_log_handler = MockLoggingHandler(logging.DEBUG)
-        logger.addHandler(cls._task_log_handler)
 
     def setUp(self):
         self.global_config = BaseGlobalConfig()
@@ -87,16 +79,8 @@ class TestCommandTask(unittest.TestCase):
                 task._handle_returncode(1, stderr)
 
 
-class TestCommandTaskWithMockPopen(unittest.TestCase):
+class TestCommandTaskWithMockPopen(MockLoggerMixin, unittest.TestCase):
     """ Run command tasks with a mocked popen """
-
-    @classmethod
-    def setUpClass(cls):
-        super(TestCommandTaskWithMockPopen, cls).setUpClass()
-        logger = logging.getLogger(cumulusci.core.tasks.__name__)
-        logger.setLevel(logging.DEBUG)
-        cls._task_log_handler = MockLoggingHandler(logging.DEBUG)
-        logger.addHandler(cls._task_log_handler)
 
     def setUp(self):
         self.global_config = BaseGlobalConfig()
